Separate properly services interface and standard service implementation
* Service interface: `MagneticFieldService.h`

* Service interface implementation: `MagneticFieldService_service.cc`

* `MagneticFieldServiceStandard.h` is deprecated because the class has
  no unique interface.

* Consequently,
  `nug4::MagneticFieldServices_MagneticFieldServiceStandard_service` is
  _not_ linkable: link to the `nug4::MagneticFieldService` `INTERFACE`
  library and utilize FHiCL and art's plugin system.

* Classify carefully `PUBLIC`, `PRIVATE`, and `INTERFACE` dependencies.
