+++ nug4/MagneticFieldServices/CMakeLists.txt
@@ -1,14 +1,17 @@
+cet_make_library(INTERFACE SOURCE MagneticFieldService.h
+  LIBRARIES INTERFACE
+  art_plugin_types::serviceDeclaration
+  nug4::MagneticField
+)
 
-cet_build_plugin( MagneticFieldServiceStandard art::service
-      LIBRARIES PRIVATE
-            messagefacility::MF_MessageLogger
-            cetlib_except::cetlib_except
-            Geant4::G4geometry
-            CLHEP::CLHEP
-            PUBLIC
-            nug4::MagneticField
-            ROOT::Core
-            ROOT::Geom )
+cet_build_plugin(MagneticFieldServiceStandard art::service
+  LIBRARIES PRIVATE
+  nug4::MagneticFieldStandard
+  nug4::MagneticFieldServices
+  art::Framework_Services_Registry
+  art::Framework_Principal
+  fhiclcpp::fhiclcpp
+)
 
 install_headers()
 install_fhicl()
