+++ nug4/MagneticFieldServices/MagneticFieldService.h
@@ -11,7 +11,6 @@
 
 #include "nug4/MagneticField/MagneticField.h"
 
-#include "art/Framework/Services/Registry/ServiceHandle.h"
 #include "art/Framework/Services/Registry/ServiceDeclarationMacros.h"
 
 namespace mag {
